+++ app/Models/Post.php
@@ -63,4 +63,14 @@ class Post extends Model
     public function  hasTag($tag_id){
         return in_array($tag_id, $this->tags->pluck('id')->toArray());
     }
+
+    public function scopeSearched($query){
+        $search = request()->query('search');
+
+        if(!$search){
+         return $query;
+        }
+
+        return $query->where('title', 'LIKE', "%{$search}%");
+    }
 }
