+++ routes/web.php
@@ -22,8 +22,8 @@ use App\Http\Controllers\WelcomeController;
 Route::get('/', [WelcomeController::class, 'index'])->name('website');
 Route::get('/blog/authors/{author}', [WelcomeController::class, 'author'])->name('blog.author');
 Route::get('/blog/post/{post}', [BlogSinglePostController::class, 'show'])->name('blog.post');
-Route::get('/blog/categories/{category}', [BlogSinglePostController::class, 'category'])->name('blog.category');
-Route::get('/blog/tags/{tag}', [BlogSinglePostController::class, 'tag'])->name('blog.tag');
+Route::get('/blog/categories/{category}', [WelcomeController::class, 'category'])->name('blog.category');
+Route::get('/blog/tags/{tag}', [WelcomeController::class, 'tag'])->name('blog.tag');
 
 Route::get('/dashboard', function () {
     return view('dashboard')->with('dashboard', true);
